New route now connects
