+++ config/routes.rb
@@ -1,8 +1,6 @@
 Rails.application.routes.draw do
-  namespace 'api' do
-    namespace 'v1' do
-      match '/organisations', to: 'organisations#create', via: [:post]
-    end
+  namespace 'identifiers' do
+    match '/ppon', to: 'organisations#create', via: [:post]
   end
   post '*unmatched_route', to: 'application#not_found'
 end
